% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volta.R
\name{get_peak2_data}
\alias{get_peak2_data}
\title{Load Peak 2 FCS files at different PMT voltages}
\usage{
get_peak2_data(
  path = ".",
  list.dirs... = list(),
  list.files... = list(),
  read.FCS... = list(),
  min_zero = FALSE,
  cv_fun = plinth::bd_rcv,
  cv_multiplier = 100,
  replace_with_na_condition = ~is.infinite(.x) | is.nan(.x) | .x < 0,
  color_nm_map = plinth::color_nm_map
)
}
\arguments{
\item{path}{Path to directory holding the Peak 2 FCS files for analysis.}

\item{list.dirs...}{An optional list of arguments to \code{\link[base:list.dirs]{base::list.dirs()}}, for finer selection control of the subdirectories of \code{path}.}

\item{list.files...}{An optional list of arguments to \code{\link[base:list.files]{base::list.files()}}, for finer selection control of Peak 2 FCS files under \code{path}.}

\item{read.FCS...}{An optional list of arguments to \code{\link[flowCore:read.FCS]{flowCore::read.FCS()}}, for reading in the Peak 2 FCS files.}

\item{min_zero}{Logical; if \code{TRUE}, shift expression values for each channel so that the minimum value is zero. The default is \code{FALSE}.}

\item{cv_fun}{Function to calculate the coefficient of variation against which PMT voltage is plotted. The default is BD Biosciences robust standard deviation (BD-rSD), \code{bd_rcv}; alternatives are \code{cv} & robust coefficient of variation (rCV), \code{rcv}.}

\item{cv_multiplier}{Numeric; the default value of 100 pushes all CVs above 1.}

\item{replace_with_na_condition}{Optional argument \code{condition} to \code{\link[naniar:replace_with_na_all]{naniar::replace_with_na_all()}} to replace e.g. negative CV values with \code{NA}.}

\item{color_nm_map}{Named vector map of color names to nm wavelength values, the default being \code{\link[plinth:color_nm_map]{plinth::color_nm_map}}.}
}
\value{
\code{\link[=get_peak2_data]{get_peak2_data()}} returns a list (named after the input FCS files) of data matrices, each of whose first column contains all test voltages, with remaining columns containing Peak 2 CV values for each voltage. Each data matrix in the list has the following custom attributes:
\item{\code{experiment_name}}{Usually the name of a directory holding a collection of per-voltage FCS files for a particular instrument.}
\item{\code{color}}{Color names for each channel.}
\item{\code{wavelength}}{Color wavelengths (nm) for each channel.}
\item{\code{raw_cv}}{The CV data matrix for this instument before removal of e.g. negatives, \code{NaN}s, & so on.}
}
\description{
Loads Peak 2 FCS files to gather up channel & voltage data for "knee" analysis.
}
\examples{
\dontrun{
library(volta)
pacman::p_load(magrittr)
x <- get_peak2_data("QC/Peak 2 and Baseline Calibrations/2020/March 2020/Data",
  list.dirs... = list(pattern = "(Waldorf 030220|Z01)$", recursive = TRUE))
}
}
