% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volta-viz.R
\name{plot_voltration_data}
\alias{plot_voltration_data}
\title{Plot Peak 2 CV vs. PMT voltages & find minimum "knee" voltage}
\usage{
plot_voltration_data(
  x,
  report_path = NULL,
  image_dir = NULL,
  save_png = FALSE,
  png... = list(),
  x_var_lab = c(PMT_voltage = "PMT Voltage"),
  y_var_lab = c(log10_CV = expression(paste(log[10], " CV"))),
  plot_series... = list(),
  points... = list(),
  xlsx_expression = NULL,
  plot_individual_channels = FALSE
)
}
\arguments{
\item{x}{List object generated by \link{get_peak2_data}}

\item{image_dir}{...}

\item{save_png}{...}

\item{plot_series...}{...}

\item{points...}{...}

\item{xlsx_expression}{...}

\item{report_dir}{[Coming soon]}

\item{trans_fun}{...}

\item{replace_with_na_condition}{...}

\item{remove_empty_cols}{...}

\item{create_smooth_variables...}{...}

\item{plot_derivative}{...}

\item{debug}{...}
}
\description{
Creates a plot of Peak 2 CV vs. PMT voltages for each channel for each instrument & finds minimum "knee" voltage for optimal resolution sensitivity. Optionally generates a summary report as an Excel workbook.
}
\examples{
\dontrun{
## There will totally be code here soon.
}
}
